Fix saving function. Re-write html/PHP of dropdown to load current value (if any) from DB.
Module rendered, but was not saving selection and, once that was fixed,  was also not displaying saved values.

Test Plan: Check that 'Focus Level' field displays/functions, is consistent with other
field options (ie'Author'), saves a meta_key of 'candela_focus_rating'
with a meta_value matching the value selected in the dropdown field.